Work on admin section started and basic login for admins complete. More work needed on this to get it to a finished status. Created admin home page as well as an affix for navigation, also created admin_managment page so new admin logins can be created.
Creation of news and courses pages with update to nav, these pages are currently blank and will be worked on in later sprints.

Modal added to header to admin login so it is accessible not matter what page the user is on.

Adding dump of database to commits from this point forward.